changed scroll onboarding screen
